Add Burning Ship fractal to help documents
help.src BurningShip added
help2.src HT_BURNINGSHIP and HF_BURNINGSHIP defined and formula given.
Also a description of the fractal is added
help5.src the name of the program ManpWIN corrected
HT_BURNINGSHIP and HF_BURNINGSHIP fixed in fractalp.cpp for Burning Ship.
