+++ hc/src/help.src
@@ -116,6 +116,7 @@
 {2.52 , 1, "Latoocarfian"}
 {2.53 , 1, "Mandelbrot Mix 4"}
 {2.54 , 1, "DivideBrot5"}
+{2.55 , 1, "Burning Ship"}
 
 {3.   , 0, Doodads\, Bells\, and Whistles, FF}
 {3.1  , 1, "Drawing Method"}
