+++ hc/src/help5.src
@@ -1225,7 +1225,7 @@ current editions of Visual Studio (as of April, 2024) is here:
 Kalles-Fraktaler is a fractal renderer with SIMD and multithreading support.
 Download from <https://fraktaler.mathr.co.uk/>
 
-ManPWin is a Windows fractal generator derived from DOS FRACTINT with
+ManpWIN is a Windows fractal generator derived from DOS FRACTINT with
 lots of additional features by Paul de Leeuw.  You can download
 the latest release from his web site: <https://www.deleeuw.com.au/>;
 Paul has collaborated with Richard Thomson to put his source code on
