+++ libid/fractalb.cpp
@@ -1073,40 +1073,6 @@ julia_z_power_bf_fractal()
     return g_bailout_bigfloat();
 }
 
-int
-burning_ship_bf_fractal()
-{
-    int degree = (int) g_params[2];
-    BFComplex big_z;
-    int saved;
-    saved = save_stack();
-
-    if (degree == 2) // Burning Ship
-    {
-        square_bf(g_tmp_sqr_x_bf, g_old_z_bf.x);
-        square_bf(g_tmp_sqr_y_bf, g_old_z_bf.y);
-        sub_bf(g_bf_tmp, g_tmp_sqr_x_bf, g_tmp_sqr_y_bf);
-        add_bf(g_new_z_bf.x, g_bf_tmp, g_parm_z_bf.x);
-        mult_bf(g_bf_tmp, g_old_z_bf.x, g_old_z_bf.y);
-        double_a_bf(g_bf_tmp);
-        abs_a_bf(g_bf_tmp);
-        sub_bf(g_new_z_bf.y, g_bf_tmp, g_parm_z_bf.y);
-    }
-    else if (degree > 2) // Burning Ship to higher power
-    {
-        big_z.x = alloc_stack(g_bf_length + 2);
-        big_z.y = alloc_stack(g_bf_length + 2);
-        abs_bf(big_z.x, g_old_z_bf.x);
-        abs_bf(big_z.y, g_old_z_bf.y);
-        neg_a_bf(big_z.y);
-        cmplx_poly_bf(&g_new_z_bf, big_z, degree);
-        add_a_bf(g_new_z_bf.x, g_parm_z_bf.x);
-        add_a_bf(g_new_z_bf.y, g_parm_z_bf.y);
-        restore_stack(saved);
-    }
-    return g_bailout_bigfloat();
-}
-
 BFComplex *cmplx_log_bf(BFComplex *t, BFComplex *s)
 {
     if (is_bf_zero(s->x) && is_bf_zero(s->y))
@@ -1174,63 +1140,6 @@ BFComplex *cmplx_div_bf(BFComplex *t, BFComplex *x, BFComplex *y)
 
     restore_stack(saved);
     return t;
-}
-
-BFComplex *cmplx_poly_bf(BFComplex *out, BFComplex xx, int degree)
-
-{
-    bf_t t, t1, t2, t3, t4;
-    int cplxsaved;
-
-    cplxsaved = save_stack();
-    t = alloc_stack(g_r_bf_length + 2);
-    t1 = alloc_stack(g_r_bf_length + 2);
-    t2 = alloc_stack(g_r_bf_length + 2);
-    t3 = alloc_stack(g_r_bf_length + 2);
-    t4 = alloc_stack(g_r_bf_length + 2);
-
-    if (degree < 0)
-        degree = 0;
-
-    copy_bf(t1, xx.x); // BigTemp1 = xt
-    copy_bf(t2, xx.y); // BigTemp2 = yt
-
-    if (degree & 1)
-    {
-        copy_bf(out->x, t1); // new.x = result real
-        copy_bf(out->y, t2); // new.y = result imag
-    }
-    else
-    {
-        int_to_bf(out->x, 1);
-        int_to_bf(out->y, 0);
-    }
-
-    degree >>= 1;
-    while (degree)
-    {
-        sub_bf(t, t1, t2);  // (xt - yt)
-        add_bf(t3, t1, t2); // (xt + yt)
-        mult_bf(t4, t, t3); // t2 = (xt + yt) * (xt - yt)
-        copy_bf(t, t2);
-        mult_bf(t3, t, t1); // yt = xt * yt
-        add_bf(t2, t3, t3); // yt = yt + yt
-        copy_bf(t1, t4);
-
-        if (degree & 1)
-        {
-            mult_bf(t, t1, out->x);  // xt * result->x
-            mult_bf(t3, t2, out->y); // yt * result->y
-            sub_bf(t4, t, t3);       // t2 = xt * result->x - yt * result->y
-            mult_bf(t, t1, out->y);  // xt * result->y
-            mult_bf(t3, t2, out->x); // yt * result->x
-            add_bf(out->y, t, t3);   // result->y = result->y * xt + yt * result->x
-            copy_bf(out->x, t4);     // result->x = t2
-        }
-        degree >>= 1;
-    }
-    restore_stack(cplxsaved);
-    return out;
 }
 
 BFComplex *cmplx_pow_bf(BFComplex *t, BFComplex *xx, BFComplex *yy)
