+++ libid/include/fractalb.h
@@ -40,12 +40,10 @@ int mandel_bf_per_pixel();
 int julia_bf_per_pixel();
 int julia_bf_fractal();
 int julia_z_power_bf_fractal();
-int burning_ship_bf_fractal();
 BFComplex *cmplx_log_bf(BFComplex *t, BFComplex *s);
 BFComplex *cmplx_mul_bf(BFComplex *t, BFComplex *x, BFComplex *y);
 BFComplex *cmplx_div_bf(BFComplex *t, BFComplex *x, BFComplex *y);
 BFComplex *cmplx_pow_bf(BFComplex *t, BFComplex *xx, BFComplex *yy);
-BFComplex *cmplx_poly_bf(BFComplex *out, BFComplex xx, int degree);
 int bn_mod_bailout();
 int bn_real_bailout();
 int bn_imag_bailout();
