+++ libid/include/fractals.h
@@ -55,10 +55,6 @@ void mandel_ref_pt(const std::complex<double> &center, std::complex<double> &z);
 void mandel_ref_pt_bf(const BFComplex &center, BFComplex &z);
 void mandel_perturb(
     const std::complex<double> &ref, std::complex<double> &delta_n, const std::complex<double> &delta0);
-void burning_ship_ref_pt(const std::complex<double> &center, std::complex<double> &z);
-void burning_ship_ref_pt_bf(const BFComplex &center, BFComplex &z);
-void burning_ship_perturb(
-    const std::complex<double> &ref, std::complex<double> &delta_n, const std::complex<double> &delta0);
 int julia_fp_per_pixel();
 int other_mandel_fp_per_pixel();
 int other_julia_fp_per_pixel();
